Changes dialect classes and their function
The dialect classes are now called

- LanguageComment : Interface of a class returning informations
                    about comments
- Language        : Interface extending LanguageComment and FileFilter.
                    Returns the dialect identifier.
- Java            : Class implementing Language
- Cpp             : Class implementing Language

The main function now fills a TreeMap with the identifiers of the
languages as the key to a object of the language. To add a new language
you only have to create a new class implementing the Language interface
and add this class to the TreeMap.

The constructor of the Changer class now asks for a language object
containing the informations about the comments and the related
FileFilter.